Reorg people + marketing domains
Move the people directory (people.ts + contacts.ts) into domains/people/ and the
public waitlist into domains/marketing/, same full-structure pattern as
notifications: tables composed into schema.ts, plain handlers in the domain, and
thin root facades preserving api.people.listPeople / api.contacts.listContacts /
api.waitlist.join.

api.d.ts diff is purely additive (new domain module entries); no callable path
or validator changed. Schema composition verified identical by the integration
suite. Backend 100 unit + 25 integration + types green.
