+++ packages/backend/convex/_generated/api.d.ts
@@ -18,10 +18,14 @@ import type * as calendar from "../calendar.js";
 import type * as calendarSync from "../calendarSync.js";
 import type * as contacts from "../contacts.js";
 import type * as crons from "../crons.js";
+import type * as domains_marketing_mutations from "../domains/marketing/mutations.js";
+import type * as domains_marketing_tables from "../domains/marketing/tables.js";
 import type * as domains_notifications_model from "../domains/notifications/model.js";
 import type * as domains_notifications_mutations from "../domains/notifications/mutations.js";
 import type * as domains_notifications_queries from "../domains/notifications/queries.js";
 import type * as domains_notifications_tables from "../domains/notifications/tables.js";
+import type * as domains_people_queries from "../domains/people/queries.js";
+import type * as domains_people_tables from "../domains/people/tables.js";
 import type * as googleSync from "../googleSync.js";
 import type * as healthCheck from "../healthCheck.js";
 import type * as http from "../http.js";
@@ -64,10 +68,14 @@ declare const fullApi: ApiFromModules<{
   calendarSync: typeof calendarSync;
   contacts: typeof contacts;
   crons: typeof crons;
+  "domains/marketing/mutations": typeof domains_marketing_mutations;
+  "domains/marketing/tables": typeof domains_marketing_tables;
   "domains/notifications/model": typeof domains_notifications_model;
   "domains/notifications/mutations": typeof domains_notifications_mutations;
   "domains/notifications/queries": typeof domains_notifications_queries;
   "domains/notifications/tables": typeof domains_notifications_tables;
+  "domains/people/queries": typeof domains_people_queries;
+  "domains/people/tables": typeof domains_people_tables;
   googleSync: typeof googleSync;
   healthCheck: typeof healthCheck;
   http: typeof http;
